fix(clippy): workspace --fix + 2 correctness errors (231->99)
- erasing_op display.rs bg_idx (0*fb_w always 0), approx_constant PI in test
- 61 file mechanical fixes via clippy --fix (unused imports, casts,
  assert_eq bool, field reassign, needless borrow, useless vec, mut)

+++ crates/mivon-emu/examples/probe_tramp.rs
@@ -41,7 +41,7 @@ fn main() {
     println!("code @pc: {}", line);
     println!(
         "console_bytes: {:02x?}",
-        cpu.out.iter().map(|b| *b as u8).collect::<Vec<_>>()
+        cpu.out.iter().map(|b| *b).collect::<Vec<_>>()
     );
     // Dump 0x9080..0x9180 (grub_bios_interrupt + prot/real trampolin).
     let mut v = Vec::with_capacity(0x100);
